Moved VM conversion code into core/vm
