+++ src/gateway/core/vm/conversion/path.go
