+++ src/gateway/core/vm/conversion/path_test.go
